src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
fastjsonschema==2.19.1
    # via -r requirements.in
genson==1.3.0
    # via -r requirements.in
jsonschema==4.22.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via wikitextparser
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via -r requirements.in
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
wcwidth==0.2.13
    # via wikitextparser
wikitextparser==0.55.13
    # via -r requirements.in
